PROTON-686, PROTON-???: Fix recent changes to compile on windows
git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/trunk@1625220 13f79535-47bb-0310-9956-ffa450edef68

+++ proton-c/src/windows/iocp.h
@@ -47,7 +47,7 @@ struct iocp_t {
   char *shared_pool_memory;
   write_result_t **shared_results;
   write_result_t **available_results;
-  int shared_available_count;
+  size_t shared_available_count;
   size_t writer_count;
   int loopback_bufsize;
   bool iocp_trace;
